requirements.txt
streamlit
crewai
pdfplumber
python-docx
python-dotenv
